% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{create_MCMC}
\alias{create_MCMC}
\title{Main wrapper function for MCMC.}
\usage{
create_MCMC(
  adjmat,
  coefEdges,
  coefTriangle,
  MCMC_interval = 1024L,
  MCMC_samplesize = 1024L,
  MCMC_burnin = 1024 * 16L,
  p_one_node_swap = 0.01,
  p_large_step = 0.01,
  p_invert = 0.01,
  lambda = 0.5,
  full_sample = FALSE,
  verbose = 0L
)
}
\description{
Main wrapper function for MCMC.
}
