% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MCMLE.R
\name{myergm_MCMLE}
\alias{myergm_MCMLE}
\title{Implement an MCMLE to estimate parameters.}
\usage{
myergm_MCMLE(
  model,
  seed = NULL,
  MCMC_interval = 1024,
  MCMC_samplesize = 1024,
  MCMC_burnin = 1024 * 16,
  p_one_node_swap = 0.01,
  p_large_step = 0.01,
  p_invert = 0.001,
  lambda = 0.5,
  verbose = 0,
  second_round = TRUE,
  ...
)
}
\arguments{
\item{model}{a model to be estimated. It only supports \code{network ~ edges + triangle}.}

\item{seed}{Seed value (integer) for the random number generator.}

\item{MCMC_interval}{Number of proposals between sampled statistics.}

\item{MCMC_samplesize}{Number of network statistics, randomly drawn from a given distribution on the set of all networks,
returned by the Metropolis-Hastings algorithm.}

\item{MCMC_burnin}{Number of proposals before any MCMC sampling is done.}

\item{p_one_node_swap}{Probabiluty of changing all links of aone node.}

\item{p_large_step}{Probability of swapping multiple links in one step.
The number of links to be swapped is determined by lambda * N.}

\item{p_invert}{Probability of inverting the adjacency matrix.
When the number of vertices is large, it is recommended to set this probaility to be zero.}

\item{lambda}{Fraction of links updated for large step.}

\item{verbose}{A logical or an integer: if this is TRUE/1, the program will print out additional information about the progress of estimation and simulation.
Higher values produce more verbosity.}

\item{second_round}{A logical: If TRUE, MCMLE is implemented twice. Otherwise, only once.}

\item{...}{Additional arguments, passed to other functions.}
}
\value{
Estimates of \code{edges} and \code{triangle} parameters
}
\description{
Implement an MCMLE to estimate parameters.
}
\examples{
# Load a network object
library(ergm)
data(florentine)

# Estimate the parameters
theta <- myergm_MCMLE(model = flomarriage ~ edges + triangle, seed = 334)

theta
}
